requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
python-dotenv>=1.0.0
langsmith
sentence-transformers>=2.2.0
torch
typing-extensions
pathlib
